Delete fix
O problema do delete dos itens residia no fato do id passado para que fosse feito a exclusão ser da posição do item no recyclerview e não o id do próprio item.

+++ app/src/main/java/br/edu/ifsp/dmo/diario/ui/listeners/DiaryEntryItemClickListener.kt
@@ -3,5 +3,5 @@ package br.edu.ifsp.dmo.diario.ui.listeners
 interface DiaryEntryItemClickListener {
     fun clickOpenDiaryEntry(position: Int)
 
-    fun clickDeleteDiaryEntry(position: Int)
+    fun clickDeleteDiaryEntry(position: Long)
 }
